refactor: Move loggers to under logging package
