+++ src/main/kotlin/org/rooftop/netx/engine/logging/LoggerFactory.kt
@@ -1,4 +1,4 @@
-package org.rooftop.netx.engine
+package org.rooftop.netx.engine.logging
 
 import ch.qos.logback.classic.Level
 import ch.qos.logback.classic.Logger
